backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
openai==0.27.8
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.22.0
